Move controllers than run in the master-controller-manager into common dir
