+++ api/cmd/master-controller-manager/controllers.go
@@ -5,14 +5,14 @@ import (
 	"fmt"
 	"time"
 
-	projectlabelsynchronizer "github.com/kubermatic/kubermatic/api/pkg/controller/project-label-synchronizer"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/rbac"
+	projectlabelsynchronizer "github.com/kubermatic/kubermatic/api/pkg/controller/master-controller-manager/project-label-synchronizer"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/master-controller-manager/rbac"
+	seedproxy "github.com/kubermatic/kubermatic/api/pkg/controller/master-controller-manager/seed-proxy"
+	seedsync "github.com/kubermatic/kubermatic/api/pkg/controller/master-controller-manager/seed-sync"
+	serviceaccount "github.com/kubermatic/kubermatic/api/pkg/controller/master-controller-manager/service-account"
+	userprojectbinding "github.com/kubermatic/kubermatic/api/pkg/controller/master-controller-manager/user-project-binding"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/master-controller-manager/usersshkeyssynchronizer"
 	seedcontrollerlifecycle "github.com/kubermatic/kubermatic/api/pkg/controller/seed-controller-lifecycle"
-	seedproxy "github.com/kubermatic/kubermatic/api/pkg/controller/seed-proxy"
-	seedsync "github.com/kubermatic/kubermatic/api/pkg/controller/seed-sync"
-	serviceaccount "github.com/kubermatic/kubermatic/api/pkg/controller/service-account"
-	userprojectbinding "github.com/kubermatic/kubermatic/api/pkg/controller/user-project-binding"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/usersshkeyssynchronizer"
 	kubermaticclientset "github.com/kubermatic/kubermatic/api/pkg/crd/client/clientset/versioned"
 	"github.com/kubermatic/kubermatic/api/pkg/crd/client/informers/externalversions"
 	kubermaticlog "github.com/kubermatic/kubermatic/api/pkg/log"
