+++ api/pkg/controller/master-controller-manager/rbac/sync_resource.go
