+++ api/pkg/controller/master-controller-manager/rbac/sync_resource_test.go
@@ -4,8 +4,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/kubermatic/kubermatic/api/pkg/controller/rbac/test"
-	fakeInformerProvider "github.com/kubermatic/kubermatic/api/pkg/controller/rbac/test/fake"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/master-controller-manager/rbac/test"
+	fakeInformerProvider "github.com/kubermatic/kubermatic/api/pkg/controller/master-controller-manager/rbac/test/fake"
 	kubermaticv1 "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
 
 	k8scorev1 "k8s.io/api/core/v1"
