+++ api/pkg/controller/user-cluster-controller-manager/rbac/mapper_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/kubermatic/kubermatic/api/pkg/controller/rbac"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/master-controller-manager/rbac"
 
 	"k8s.io/apimachinery/pkg/api/equality"
 )
